+++ Bird-Person-Game/Source/BPG/Utils/Animation.cpp
@@ -3,7 +3,7 @@
 namespace Utils
 {
 
-	Animation::Animation(const sf::Texture & texture, size_t cols, size_t rows, size_t startX, size_t lengthX, size_t startY, size_t lengthY) :
+	Animation::Animation(const sf::Texture & texture, uint32_t cols, uint32_t rows, uint32_t startX, uint32_t lengthX, uint32_t startY, uint32_t lengthY) :
 		holdTime(sf::milliseconds(100)),
 		index(0u),
 		texture(texture),
@@ -13,9 +13,9 @@ namespace Utils
 		const uint32_t tileWidth = size.x / cols;
 		const uint32_t tileHeight = size.y / rows;
 
-		for (size_t y = startY; y < startY + lengthY; y++)
+		for (uint32_t y = startY; y < startY + lengthY; y++)
 		{
-			for (size_t x = startX; x < startX + lengthX; x++)
+			for (uint32_t x = startX; x < startX + lengthX; x++)
 			{
 				const uint32_t xpos = x * tileWidth;
 				const uint32_t ypos = y * tileHeight;
